+++ app/helpers/service/header_helper.rb
@@ -9,17 +9,6 @@ module Service::HeaderHelper
     [provider_links, status]
   end
 
-  def resource_link(service)
-    if service.pid.present?
-      "#{Mp::Application.config
-                        .providers_dashboard_url}/resource-dashboard/#{service.pid
-                                                                              .split(".")
-                                                                              .first}/#{service.pid}/stats"
-    else
-      service_path(service)
-    end
-  end
-
   def my_providers_link
     "#{Mp::Application.config.providers_dashboard_url}/provider/my"
   end